requirements.txt
absl-py==1.0.0
attrs==21.4.0
comtypes==1.1.11
cvzone==1.5.6
cycler==0.11.0
fonttools==4.29.1
imageio==2.16.1
imutils==0.5.4
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.9.1
networkx==2.7.1
numpy==1.22.2
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
protobuf==3.20.0
pygrabber==0.1
pyparsing==3.0.7
pyperclip==1.8.2
pystray==0.19.3
python-dateutil==2.8.2
pyvirtualcam==0.9.1
PyWavelets==1.3.0
pyzbar==0.1.8
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
tifffile==2022.3.25
